+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ ChangeLog
 0.2 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Only try to update the file if something changed
 
 
 0.1 (2019-11-07)