+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ ChangeLog
 
 - Only try to update the file if something changed
 - Add ``--check`` option to check if all the files are up-to-date with their generators
+- Improve diff shown with ``--verbose`` option
 
 
 0.1 (2019-11-07)